Change timeStamp to Int64

+++ business-chat-app/business-chat-app/Model/Chat.swift
@@ -14,7 +14,7 @@ class Chat {
     private var _memberCount: String
     private var _members: [String:Bool]
     private var _chatKey: String
-    private var _lastMessage: Double
+    private var _lastMessage: Int64
     
     var chatName: String {
         return _chatName
@@ -29,11 +29,11 @@ class Chat {
     var key: String {
         return _chatKey
     }
-  var lastMessage: Double {
+  var lastMessage: Int64 {
     return _lastMessage
   }
     
-  init(name: String, members: [String:Bool], chatKey: String, memberCount: String, lastMessage: Double) {
+  init(name: String, members: [String:Bool], chatKey: String, memberCount: String, lastMessage: Int64) {
         self._chatName = name
         self._members = members
         self._chatKey = chatKey
