+++ business-chat-app/business-chat-app/Services/DataServices.swift
@@ -94,8 +94,8 @@ class Services {
       con.setValue(true)
       var date = Date()
       
-      let currentDate = Double((date.millisecondsSince1970))
-       date = Date(milliseconds: Int(currentDate))
+      let currentDate = Int64((date.millisecondsSince1970))
+       date = Date(milliseconds: Int64(currentDate))
       // when I disconnect, update the last time I was seen online
       lastOnlineRef.onDisconnectSetValue(currentDate)
       offlineAfterDisconnect.onDisconnectSetValue("offline")
